requirements.txt
Django==4.2.18
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dj-rest-auth[with_social]==5.0.2
django-allauth==0.61.1
celery==5.4.0
redis==5.0.4
requests==2.31.0
python-dotenv==1.0.1
gunicorn==22.0.0 # Recommended for production